# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Skill Hunters — contrat initial L0",
    "version": "0.4.0",
    "description": "Contrat MVP. x-implementation-status distingue les opérations livrées dans le premier incrément identité des parcours encore planifiés."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/imports": {
      "post": {
        "operationId": "createImport",
        "tags": [
          "Documents"
        ],
        "summary": "Déposer un document en quarantaine privée",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "Réception persistée, accessible uniquement aux membres du vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImportRequest"
              },
              "encoding": {
                "basis": {
                  "contentType": "application/json"
                }
              }
            }
          }
        },
        "x-implementation-status": "implemented",
        "description": "Limite 15 Mio, un fichier. Idempotence par acteur/vivier/clé ; conflit si empreinte ou métadonnées diffèrent. Aucun parsing métier. Quotas initiaux : 1 000 fichiers et 3 000 Mio par acteur, réservations de lots incluses ; quatre réceptions actives par instance. DOC/PDF nécessitent encore qualification isolée, antivirus et limite de pages. Le préfixe de clé batch- est réservé au service de lots. Pour plusieurs fichiers utiliser /import-batches."
      }
    },
    "/imports/{id}": {
      "get": {
        "operationId": "getImport",
        "tags": [
          "Documents"
        ],
        "summary": "Suivre les résultats par fichier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Réception persistée, accessible uniquement aux membres du vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "x-implementation-status": "implemented"
      }
    },
...
    "/pools/{poolId}/imports": {
      "get": {
        "operationId": "listPoolImports",
        "tags": [
          "Documents"
        ],
        "summary": "Lister les réceptions autorisées par curseur",
        "parameters": [
          {
            "name": "poolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Curseur nextCursor de la page précédente ; ordre date/id décroissant."
          }
        ],
        "x-implementation-status": "implemented",
        "responses": {
          "200": {
            "description": "Résultat filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/pools/{poolId}/import-policies": {
      "get": {
        "operationId": "listImportPolicies",
        "tags": [
          "Documents"
        ],
        "summary": "Lister les cadres de dépôt configurés",
        "parameters": [
          {
            "name": "poolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-implementation-status": "implemented",
        "responses": {
          "200": {
            "description": "Résultat filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportPolicy"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/import-batches": {
      "post": {
        "operationId": "createImportBatch",
        "summary": "Réserver un manifeste de 1 à 100 fichiers",
        "tags": [
          "Documents"
        ],
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "$ref": "#/components/parameters/CsrfToken"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "responses": {
          "202": {
            "description": "État durable de la réception, filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchManifest"
              }
            }
          }
        },
        "description": "300 Mio par lot. Idempotence acteur/vivier/clé/empreinte du manifeste ordonné. Politique figée. Un rejet individuel est conservé sans annuler le lot. Les quotas incluent les réservations."
      }
    },
    "/import-batches/{id}": {
      "get": {
        "operationId": "getImportBatch",
        "summary": "Consulter progression, tentatives et résultats du lot",
        "tags": [
          "Documents"
        ],
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "État durable de la réception, filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/pools/{poolId}/import-batches": {
      "get": {
        "operationId": "listImportBatches",
        "summary": "Historique paginé des lots du vivier",
        "tags": [
          "Documents"
        ],
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "poolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "État durable de la réception, filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatchPage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        }
      }
    },
    "/import-batches/{batchId}/items/{itemId}/content": {
      "put": {
        "operationId": "uploadBatchItem",
        "summary": "Envoyer ou reprendre un fichier réservé",
        "tags": [
          "Documents"
        ],
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "batchId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "État durable de la réception, filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "description": "Auteur habilité uniquement. Identifiant de fichier stable ; nom/taille/SHA-256 vérifiés. Cinq tentatives, bail deux minutes, deux transferts actifs par auteur. Un résultat métier de rejet ou de panne reprenable est dans items[].state/errorCode avec HTTP 200 ; 200 ne signifie pas que le fichier est accepté."
      }
    },
    "/imports/{id}/content": {
      "put": {
        "operationId": "resumeIndividualImport",
        "summary": "Reprendre une ancienne réception individuelle avec son original",
        "tags": [
          "Documents"
        ],
        "x-implementation-status": "implemented",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "$ref": "#/components/parameters/CsrfToken"
          }
        ],
        "responses": {
          "200": {
            "description": "État durable de la réception, filtré par habilitation au vivier.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportBatch"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Problem"
          },
          "401": {
            "$ref": "#/components/responses/Problem"
          },
          "403": {
            "$ref": "#/components/responses/Problem"
          },
          "404": {
            "$ref": "#/components/responses/Problem"
          },
          "409": {
            "$ref": "#/components/responses/Problem"
          },
          "413": {
            "$ref": "#/components/responses/Problem"
          },
          "415": {
            "$ref": "#/components/responses/Problem"
          },
          "429": {
            "$ref": "#/components/responses/Problem"
          },
          "500": {
            "$ref": "#/components/responses/Problem"
          },
          "503": {
            "$ref": "#/components/responses/Problem"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "SH_SESSION"
      }
    },
    "parameters": {
      "CsrfToken": {
        "name": "X-CSRF-Token",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 8,
          "maxLength": 128,
          "pattern": "^[A-Za-z0-9_-]{8,128}$"
        }
      }
    },
    "responses": {
      "Problem": {
        "description": "Erreur structurée ; aucune donnée inaccessible dans le détail.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "correlationId"
        ]
      },
...
      "ProcessingBasis": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "source": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "purpose": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "basisCode": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "noticeVersion": {
            "type": "string",
            "maxLength": 80
          },
          "retentionPolicyId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "source",
          "purpose",
          "basisCode",
          "retentionPolicyId"
        ]
      },
      "ImportRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "poolId": {
            "type": "string",
            "format": "uuid"
          },
          "basis": {
            "$ref": "#/components/schemas/ProcessingBasis"
          },
          "files": {
            "type": "array",
            "minItems": 1,
            "maxItems": 1,
            "items": {
              "type": "string",
              "format": "binary"
            }
          }
        },
        "required": [
          "poolId",
          "basis",
          "files"
        ],
        "description": "SH-03 : un seul fichier par requête ; le lot SH-04 reste planifié. Partie basis de type application/json."
      },
      "ImportItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "state": {
            "type": "string",
            "enum": [
              "RECEIVING",
              "QUARANTINED",
              "SCANNING",
              "EXTRACTING",
              "REVIEW_REQUIRED",
              "VALIDATED",
              "INDEXED",
              "REJECTED_FORMAT",
              "BLOCKED_SECURITY",
              "RETRYABLE_FAILURE",
              "FAILED",
              "DELETED"
            ]
          },
          "errorCode": {
            "type": "string"
          },
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "documentVersionId": {
            "type": "string",
            "format": "uuid"
          },
          "mediaType": {
            "type": "string"
          },
          "byteSize": {
            "type": "integer",
            "minimum": 1,
            "maximum": 15728640
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        },
        "required": [
          "id",
          "fileName",
          "state"
        ]
      },
      "ImportBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportItem"
            }
          },
          "poolId": {
            "type": "string",
            "format": "uuid"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "validationVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "items"
        ],
        "description": "Réception individuelle SH-03. Aucun objet S3 ni URL de téléchargement exposé. QUARANTINED signifie réception réussie, pas antivirus réussi."
      },
...
      "ImportPolicy": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "basisCode": {
            "type": "string"
          },
          "retentionDays": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "id",
          "label",
          "purpose",
          "basisCode",
          "retentionDays"
        ]
      },
      "ImportPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportBatch"
            },
            "maxItems": 20
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "BatchFileSpec": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "fileName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "byteSize": {
            "type": "integer",
            "minimum": 0,
            "maximum": 314572800
          },
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          }
        },
        "required": [
          "fileName",
          "byteSize",
          "sha256"
        ]
      },
      "BatchManifest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "poolId": {
            "type": "string",
            "format": "uuid"
          },
          "basis": {
            "$ref": "#/components/schemas/ProcessingBasis"
          },
          "files": {
            "type": "array",
            "minItems": 1,
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/BatchFileSpec"
            }
          }
        },
        "required": [
          "poolId",
          "basis",
          "files"
        ]
      },
      "BatchUploadItem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "byteSize": {
            "type": "integer"
          },
          "sha256": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "enum": [
              "WAITING_UPLOAD",
              "UPLOADING",
              "QUARANTINED",
              "REJECTED_FORMAT",
              "RETRYABLE_FAILURE",
              "FAILED"
            ]
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ]
          },
          "attempts": {
            "type": "integer",
            "minimum": 0,
            "maximum": 5
          },
          "retryAfter": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "importId": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          }
        },
        "required": [
          "id",
          "fileName",
          "byteSize",
          "sha256",
          "state",
          "errorCode",
          "attempts",
          "retryAfter",
          "importId"
        ]
      },
      "UploadBatch": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "poolId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "canUpload": {
            "type": "boolean"
          },
          "completedFiles": {
            "type": "integer"
          },
          "totalFiles": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "$ref": "#/components/schemas/BatchUploadItem"
            }
          }
        },
        "required": [
          "id",
          "poolId",
          "createdAt",
          "expiresAt",
          "canUpload",
          "completedFiles",
          "totalFiles",
          "items"
        ]
      },
      "UploadBatchSummary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "completedFiles": {
            "type": "integer"
          },
          "totalFiles": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "createdAt",
          "completedFiles",
          "totalFiles"
        ]
      },
      "UploadBatchPage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "maxItems": 20,
            "items": {
              "$ref": "#/components/schemas/UploadBatchSummary"
            }
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      }
    }
  }
}